Add 'phoenix-core` to a workspace called phoenix
Resolves #171
