+++ .gitignore
@@ -1,4 +1,2 @@
-/target
-**/*.rs.bk
-/src/main.rs
+**/target
 Cargo.lock
