+++ circuits/CHANGELOG.md
@@ -1,0 +1,19 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+
+- Add `phoenix-circuits` as a workspace member of `phoenix` [#171]
+- Add elgamal encryption and decryption gadgets [#171]
+
+<!-- ISSUES -->
+[#171]: https://github.com/dusk-network/phoenix/issues/171
+
+<!-- VERSIONS -->
+[Unreleased]: https://github.com/dusk-network/phoenix/compare/v0.27.0...HEAD
