+++ core/README.md
@@ -1,0 +1,5 @@
+# Phoenix Core
+
+Phoenix is the transaction model used by Dusk, an open-source public blockchain with a UTXO-based architecture that allows for the execution of obfuscated transactions and confidential smart contracts.
+
+This library contains the cryptographic primitives, types and functionalities used in phoenix, except for the zero-knowledge circuits.
