+++ core/src/note.rs
