+++ core/tests/note_test.rs
